+++ fugue/src/main/java/com/atlassian/fugue/Iterators.java
@@ -38,7 +38,7 @@ class Iterators {
    *
    * @since 3.0
    */
-  static <A> boolean addAll(Collection<A> collectionToModify, Iterator<? extends A> iterator) {
+  static <A> boolean addAll(final Collection<A> collectionToModify, final Iterator<? extends A> iterator) {
     requireNonNull(collectionToModify);
     requireNonNull(iterator);
     boolean wasModified = false;
@@ -49,19 +49,24 @@ class Iterators {
   }
 
   /**
-   * Wrap an iterator to add support for the peek operation
+   * Wrap an iterator to add support for the peek operation. Do not maintain a
+   * reference to the iterator passed in as a parameter. Iterators are mutable
+   * by nature and this function makes no promises about how or when the input
+   * iterator will be mutated.
+   *
    * @param iterator iterator that may not support peek, must not be null
    * @param <A> element type
-   * @return iterator that can return the next element without removing it from the iterator
+   * @return an iterator for which {@code peek} will return Iterator#next
+   * without removing the value from the iterator
    *
    * @since 3.0
    */
-  static <A> Iterators.Peeking<A> peekingIterator(java.util.Iterator<? extends A> iterator) {
+  static <A> Iterators.Peeking<A> peekingIterator(final Iterator<? extends A> iterator) {
     if (iterator instanceof PeekingImpl) {
       // Safe to cast <? extends T> to <T> because PeekingImpl only uses T
       // covariantly (and cannot be subclassed to add non-covariant uses).
       @SuppressWarnings("unchecked")
-      PeekingImpl<A> peeking = (PeekingImpl<A>) iterator;
+      final PeekingImpl<A> peeking = (PeekingImpl<A>) iterator;
       return peeking;
     }
     return new PeekingImpl<>(iterator);
@@ -76,7 +81,7 @@ class Iterators {
     private boolean hasPeeked;
     private A peekedElement;
 
-    public PeekingImpl(java.util.Iterator<? extends A> iterator) {
+    public PeekingImpl(final Iterator<? extends A> iterator) {
       this.iterator = requireNonNull(iterator);
     }
 
@@ -88,7 +93,7 @@ class Iterators {
       if (!hasPeeked) {
         return iterator.next();
       }
-      A result = peekedElement;
+      final A result = peekedElement;
       hasPeeked = false;
       peekedElement = null;
       return result;